实现SimpleFloat解码器
