+++ Mozi.IoT.CoAP/Mozi.IoT.CoAP.nuspec
@@ -8,7 +8,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<license type="expression">MIT</license>
 		<projectUrl>https://gitee.com/myui_admin/mozi</projectUrl>
-		<description>CoAP客户端调试工具，命令行调试工具,服务端请搜索Mozi.IoT.Server</description>
+		<description>CoAP客户端调试工具，命令行调试工具，服务端请搜索Mozi.IoT.Server</description>
 		<releaseNotes>CoAP客户端调试工具</releaseNotes>
 		<copyright>Copyright © Jason 2022</copyright>
 		<tags>IoT Client 调试工具</tags>
